+++ core/src/test/java/hudson/model/ActionableTest.java
@@ -24,13 +24,18 @@
 
 package hudson.model;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
 
 import java.util.Collections;
+import java.util.List;
+
+import org.junit.Assert;
 import org.junit.Test;
+import org.jvnet.hudson.test.Issue;
 
 public class ActionableTest {
 
@@ -46,6 +51,56 @@ public class ActionableTest {
         CauseAction a3 = new CauseAction();
         thing.replaceAction(a3);
         assertEquals(Arrays.asList(a2, a3), thing.getActions());
+    }
+
+    static class ActionableOverride extends Actionable {
+        ArrayList<Action> specialActions = new ArrayList<Action>();
+
+        @Override
+        public String getDisplayName() {
+            return "nope";
+        }
+
+        @Override
+        public String getSearchUrl() {
+            return "morenope";
+        }
+
+        @Override
+        public List<Action> getActions() {
+            return specialActions;
+        }
+    }
+
+    @SuppressWarnings("deprecation")
+    @Issue("JENKINS-39555")
+    @Test
+    public void testExtensionOverrides() throws Exception {
+        ActionableOverride myOverridden = new ActionableOverride();
+        InvisibleAction invis = new InvisibleAction() {
+        };
+        myOverridden.addAction(invis);
+        Assert.assertArrayEquals(new Object[]{invis}, myOverridden.specialActions.toArray());
+        Assert.assertArrayEquals(new Object[]{invis}, myOverridden.getActions().toArray());
+
+        myOverridden.getActions().remove(invis);
+        Assert.assertArrayEquals(new Object[]{}, myOverridden.specialActions.toArray());
+        Assert.assertArrayEquals(new Object[]{}, myOverridden.getActions().toArray());
+
+        myOverridden.addAction(invis);
+        myOverridden.removeAction(invis);
+        Assert.assertArrayEquals(new Object[]{}, myOverridden.specialActions.toArray());
+        Assert.assertArrayEquals(new Object[]{}, myOverridden.getActions().toArray());
+
+        InvisibleAction invis2 = new InvisibleAction() {};
+        myOverridden.addOrReplaceAction(invis2);
+        Assert.assertArrayEquals(new Object[]{invis2}, myOverridden.specialActions.toArray());
+        Assert.assertArrayEquals(new Object[]{invis2}, myOverridden.getActions().toArray());
+
+        myOverridden.addOrReplaceAction(invis);
+        myOverridden.addOrReplaceAction(invis);
+        Assert.assertArrayEquals(new Object[]{invis2, invis}, myOverridden.specialActions.toArray());
+        Assert.assertArrayEquals(new Object[]{invis2, invis}, myOverridden.getActions().toArray());
     }
 
     @SuppressWarnings("deprecation")
